Fixed errors involving running Isaac Gym on CPU
When specifying in the Isaac Gym configs whether or not to run on GPU,
it will run on GPU anyway if it detects one. I've inserted an extra
piece of logic so that Isaac Gym can detect if cuda is available. Now,
Isaac Gym can be forced to CPU by simply changing all instances of
`torch.cuda.is_available()` to `not torch.cuda.is_available()`.