Improved SAC testing
- A large `mini_batch_size` should be used, since sample generation
happens so fast and across so many different environments, that we don't
have to worry as much about high correlations between samples.
- The agent should only start training once the buffer is at max size.
I've realized that this helps prevent the agent from getting trapped in
local optima early on (especially with Ant), when it would be training
on highly correlated samples without this fix.